wsig-server/requirements.txt
Flask
pymongo
flask-cors
flask
python-dotenv
dnspython
requests